CAN FD implementation
Add a flag to CAN message definitions which set the message as using FD
protocol if true.

Use a new generated file with the new FD flag field on the message
definitions.

Change BCM socket "struct" using an union to store the CAN frames either
using the FD struct or the classic non FD struct. A BCM socket can only
one frame type once configured.

Use as much as possible the "struct canfd_frame" in the binding and only
make a difference before writing or reading the socket. From a memory
point of view both struct are identical and only the last member differ
and could hold more data with messages of 64 bytes long. So the
canfd_frame is compatible with the can_frame and can be differentiated
by a flag set in the can_id member.

Remove now unused code processing can_frame.

Keep the diagnostic manager using the classic CAN frame.

Set the maximum number of frames that a BCM socket can handle to 257.

Bug-AGL: SPEC-1980

Change-Id: Ifcc041281ea6745fc25cbd384743761f4446f489
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>