diag-mngr: Fix sending and change init
This commit fix the wrong bcm ival1 value set because the struct
hasn't been initialized so it use a wrong value that can not be
handled by the socket.

Initialize diagnotstic manager depending on JSON configuration file

This avoid to request engine_speed signal when there isn't any diagnostic_bus
configured in the JSON file.

Change-Id: I98145a6e20c29a644c856d818f104c6bd8069f9f
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>