Fix: pushing event with a diagnostic signal
Invoke jsonify_vehicle with no signal because the we are pushing a diagnostic
signal, which aren't stored in a same object, cause a segfault because signal
part of the vehicle message is empty. Now jsoniy_vehicle handle correctly the
case and just try to get the signal if only is there is a signal to parse.
Else it just rely on the data holded by the vehicle_message struct.

Change-Id: I05e0e3ab29f99d67965317358e358c6eb80168d7
Signed-off-by: Romain Forlot <romain.forlot@iot.bzh>