[verbname] allow every char for signal verbs name
Some PGN name contain the same letters but different digits in their names.
The canbus binding replaced every char excepting letters by a "_".
This implies multiple verbs definition with the same name, and so the canbus crashed.

Because the application framework allows any char to be in verb name, the canbus accept it too.

Signed-off-by: Marc-Antoine Riou <marc-antoine.riou@iot.bzh>