Fix diagnostic message frequency & enforce minimum freq
Old code was using the first diagnostic message's frequency for all
diagnostic messages. In the case of the vcar plugin, that means all
messages were set at 5Hz, using 6.5KBps of CAN bus bandwidth. Fixing
this more than halved bandwidth usage.

A frequency of 0Hz has kind of an undefined behavior, sometimes
saturating the bus, sometimes not behaving strangelly. A minimum
frequency of 0.05Hz (20 minutes period) is enforced for messages
with a specified frequency of 0Hz or no specified frequency.